Move the kotlin code to the kotlin folder of the service module #93
