+++ service/src/main/kotlin/com/wansensoft/service/role/impl/SysRoleServiceImpl.kt
@@ -10,7 +10,7 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package com.wansensoft.service.role
+package com.wansensoft.service.role.impl
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page
@@ -22,6 +22,8 @@ import com.wansensoft.entities.role.SysRole
 import com.wansensoft.entities.role.SysRoleMenuRel
 import com.wansensoft.mappers.role.SysRoleMapper
 import com.wansensoft.mappers.role.SysRoleMenuRelMapper
+import com.wansensoft.service.role.SysRoleMenuRelService
+import com.wansensoft.service.role.SysRoleService
 import com.wansensoft.utils.SnowflakeIdUtil
 import com.wansensoft.utils.constants.CommonConstants
 import com.wansensoft.utils.enums.BaseCodeEnum
